Create FeatureFetcher class
