+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/importer/steps/Download.kt
@@ -13,15 +13,14 @@ import io.ktor.http.HttpMethod
 import kotlinx.coroutines.runBlocking
 import org.slf4j.Logger
 
-class Download(private val logger: Logger) : PipelineStep<Unit, List<LbeAcceptingStore>>() {
+class Download(private val logger: Logger, private val httpClient: HttpClient) : PipelineStep<Unit, List<LbeAcceptingStore>>() {
 
     override fun execute(input: Unit): List<LbeAcceptingStore> {
         try {
             val url = System.getProperty("app.import.xml")
 
-            val client = HttpClient()
             val response = runBlocking {
-                client.request<String> {
+                httpClient.request<String> {
                     url(url)
                     method = HttpMethod.Get
                 }
